requirements-core.txt
# CareerTrojan — Core Requirements
# Merged and consolidated for Python 3.10+ compatibility
# Includes core functionality + AI Fusion Engine essentials

# ------------------------
# 🖼️ Core Backend & Frontend
# ------------------------
streamlit>=1.32.2
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.4
pydantic-settings>=2.3.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
httpx>=0.28.0
requests>=2.32.0
python-dateutil>=2.8.2
protobuf>=4.25.3,<5.0.0  # Streamlit compatibility fix
grpcio-status>=1.56.0
qrcode[pil]>=7.4.2

# ------------------------
# �️ Streamlit UI Components
# ------------------------
streamlit-extras>=0.4.0
streamlit-tags>=1.2.8
streamlit-option-menu>=0.3.13
streamlit-toggle-switch>=1.0.2
streamlit-ace>=0.1.1
streamlit-aggrid>=1.0.5
streamlit-authenticator>=0.3.0
streamlit-autorefresh>=1.0.0
zxcvbn>=4.4.28
plotly>=5.22.0  # Interactive visualizations
altair>=4.2.0   # Statistical visualizations

# ------------------------
# �📄 Document Parsing (stable wheels for Python 3.10)
# ------------------------
python-docx>=0.8.11
docx2txt>=0.8
PyPDF2>=3.0.1
pdfminer.six>=20250506  # Updated for pdfplumber compatibility
python-pptx>=0.6.23
openpyxl>=3.1.5
striprtf>=0.0.16
extract-msg>=0.23.1
fpdf2>=2.8.0  # Updated to fpdf2 for better compatibility
lxml>=4.9.4
beautifulsoup4>=4.12.2
pdfplumber>=0.11.0  # Additional PDF processing
chardet>=5.2.0  # Character encoding detection
pdf2image>=1.17.0  # PDF to image conversion (requires poppler)
python-magic-bin>=0.4.14; sys_platform == "win32"  # File type detection Windows
python-magic>=0.4.27; sys_platform != "win32"  # File type detection Unix
pypandoc>=1.15  # Universal document conversion (requires pandoc)
xmltodict>=0.13.0  # XML to dict conversion
pyorc>=0.9.0  # ORC file support
fastavro>=1.9.8  # Avro file support
pyarrow>=18.0.0  # Parquet and other columnar formats
unidecode>=1.3.8  # ASCII transliterations
PyMuPDF>=1.24.0  # Provides `fitz` import for advanced PDF processing
# textract removed due to packaging incompatibilities on modern pip; enable only with pip<24.1 if strictly required

# ⚠️ SYSTEM DEPENDENCIES 
# ✅ INSTALLED in c:\CareerTrojan\tools\bin\:
# - antiword.exe (v0.37) - Converts MS Word .doc files to text
# - pdftotext.exe (v24.08.0) - Extracts text from PDF files  
# - pdftohtml.exe (v24.08.0) - Converts PDF to HTML
# - pdftoppm.exe (v24.08.0) - Converts PDF to images
# - Other poppler utilities for PDF processing

# ------------------------
# 🤖 AI Core (light)
# ------------------------
langchain>=0.3.0
openai>=1.50.0    # Updated to modern SDK for better compatibility

# ------------------------
# 🌐 HTTP & Networking
# ------------------------
aiohttp>=3.12.0  # Async HTTP client/server
urllib3>=2.5.0  # HTTP library with connection pooling
certifi>=2025.8.3  # Root certificates
requests-oauthlib>=2.0.0  # OAuth for requests

# ------------------------
# � Email Integrations (Klaviyo, SendGrid, Resend)
# ------------------------
klaviyo-api>=12.0.0  # Klaviyo email marketing SDK
sendgrid>=6.11.0  # SendGrid transactional email
resend>=2.0.0  # Resend email API

# ------------------------
# �🔐 Security & Cryptography
# ------------------------
cryptography>=46.0.0
bcrypt>=4.2.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0

# ------------------------
# 🛠️ Database & Caching
# ------------------------
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=5.0.4  # In-memory caching
pymongo>=4.10.0  # MongoDB client
motor>=3.7.0  # Async MongoDB client

# ------------------------
# 🔤 Encoding & Validation
# ------------------------
charset-normalizer>=3.0.0
email-validator>=2.0.0  # Email validation
validators>=0.35.0  # URL and data validation
phonenumbers>=8.13.48  # Phone number parsing
bleach>=6.2.0  # HTML sanitization
markupsafe>=3.0.2  # Safe string handling

# ------------------------
# ✅ Utilities & Tools
# ------------------------
PyYAML>=6.0.1
rapidfuzz>=3.6.1
matplotlib>=3.7.0
seaborn>=0.13.2     # Statistical plotting
dateparser>=1.2.0
sqlite-utils>=3.36.0
tabulate>=0.9.0
typer>=0.12.0
psutil>=5.9.0
diskcache>=5.6.0
concurrent-log-handler>=0.9.25
prometheus-client>=0.20.0
structlog>=24.1.0
loguru>=0.7.2  # Enhanced logging
click>=8.1.0           # CLI framework
python-json-logger>=2.0.7  # JSON structured logging
rich>=13.9.0           # Rich text and beautiful formatting
tqdm>=4.66.0           # Progress bars
colorama>=0.4.6        # Cross-platform colored terminal text
pathlib2>=2.3.7        # Enhanced path handling
more-itertools>=10.8.0 # Additional iterator tools

# ------------------------
# ⚙️ Configuration & Serialization
# ------------------------
toml>=0.10.2  # TOML configuration files
tomli>=2.0.0; python_version<"3.11"  # TOML parsing for older Python
tomli-w>=1.1.0  # TOML writing
configparser>=7.1.0  # INI file parsing
json5>=0.9.28  # JSON5 format support
orjson>=3.11.3  # Fast JSON parsing
ujson>=5.11.0  # Ultra-fast JSON encoder/decoder
msgpack>=1.1.0  # MessagePack serialization
pickle5>=0.0.12; python_version<"3.8"  # Enhanced pickle

# ------------------------
# 📁 File System & Compression
# ------------------------
pathspec>=0.12.1  # Path specification matching
watchdog>=6.0.0  # File system monitoring
send2trash>=1.8.3  # Safe file deletion
zstandard>=0.25.0  # Fast compression
py7zr>=0.22.0  # 7-Zip archive support

# ------------------------
# 🔢 Math & Scientific Computing
# ------------------------
numpy>=1.26.0  # Fundamental package for scientific computing
scipy>=1.15.0  # Scientific computing library
sympy>=1.14.0  # Symbolic mathematics

# ------------------------
# ⚡ Async & Concurrency
# ------------------------
aiofiles>=24.1.0  # Async file operations
anyio>=4.10.0  # High level async library
async-timeout>=5.0.1  # Timeout handling for asyncio
asyncio-mqtt>=0.17.2  # MQTT client for asyncio
greenlet>=3.2.4  # Lightweight in-process concurrent programming

# ------------------------
# 🧪 Testing Essentials (Core Only)
# ------------------------
# Using stdlib unittest/mock/doctest — no pip packages required

# ------------------------
# 📚 NLP (base)
# ------------------------
spacy>=3.6.0
textblob>=0.17.1
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
nltk>=3.8.1
regex>=2024.11.6  # Enhanced regular expressions
fuzzywuzzy>=0.18.0  # Fuzzy string matching
python-Levenshtein>=0.25.0  # Fast string similarity
langdetect>=1.0.9  # Language detection

# ------------------------
# 📊 Data Processing Core
# ------------------------
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.5  # Excel file support (already above, but critical)
xlrd>=2.0.1  # Read Excel files
xlwt>=1.3.0  # Write Excel files
csvkit>=2.0.0  # CSV utilities
jsonlines>=4.0.0  # JSONL format support
ijson>=3.3.0  # Incremental JSON parsing

# ------------------------
# 🧠 AI Fusion Engine (Core Components)
# ------------------------
# Essential packages for Fuzzy Logic + Bayesian + LLM integration
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0  # Already above, ensuring compatibility
scikit-fuzzy>=0.4.2
pgmpy>=0.1.19
networkx>=2.6.0
scikit-learn>=1.0.0
apscheduler>=3.9.0
psutil>=5.9.0
statsmodels>=0.13.0
arrow>=1.2.0
memory-profiler>=0.60.0
